So i hope this works, this is our game so far.
Not too impressive haha, but it's a start. If you left early, we fixed
the big friction problem we were working on, feel free to play around
with this and if you do something cool post it back up, obviously we
still have a lot of physics things to work on, I added a little bit of a
bounce mechanic but that need some work too. Also keep working on
concept art, level designs, music, and any other ideas!