New version
this is pretty much what we had gotten so far except translated into the
GML script code. you'll see that the sliding now works PERFECTLY and we
can even see the value of hspeed printed on the screen. However, couldnt
get the horizontal collision to work yet :p still working on it, but
give it a shot if you have an idea!