Hi! Newest version
I made some more interesting blocks that look more neon blue instead of
like the tile to your bathroom. Also when you place blocks down they are
the red version of the new blocks. Im trying to work on changing the
color of the blocks so if the blocks all change to red dont freak out!