Storyboard for tutorial
Not colored so sorry about that, didn't have the proper tools back home.
But this should give an idea for the artist to polish.  The lines on the
trees are meant to look like candy canes and the leaves pink and blue
cotton candy.  The grass liquorice.  Hope this is a good start!