New Levels!
So these a some nice ones.  Things may be confusing give me a call or
text if there is confusion.  To clear up some, because I had minimal
objects to work with I had to use some as placeholders for things.  Red
blocks are recharging packs.  The blocks with the black outlines are
either switches that do something or special treasure.  The ones with
the down error are something bad.  Either laser shooting guys or spikes.
Anyways have a look. :)